Move getThemeFilterTermFromString
