+++ client/my-sites/themes/validate-filters.js
@@ -1,16 +1,14 @@
 /**
  * External dependencies
  */
-
 import page from 'page';
 
 /**
  * Internal dependencies
  */
-import { getThemeFilterStringFromTerm } from 'state/themes/selectors';
+import { getThemeFilterStringFromTerm, getThemeFilterTermFromString } from 'state/themes/selectors';
 
 import getThemeFilterTerm from 'state/selectors/get-theme-filter-term';
-import getThemeFilterTermFromString from 'state/selectors/get-theme-filter-term-from-string';
 import isValidThemeFilterTerm from 'state/selectors/is-valid-theme-filter-term';
 
 // Reorder and remove invalid filters to redirect to canonical URL
