+++ client/state/selectors/get-theme-filter-to-term-table.js
@@ -8,7 +8,7 @@ import { forIn, keys, mapValues } from 'lodash';
  */
 import createSelector from 'lib/create-selector';
 import getThemeFilters from 'state/selectors/get-theme-filters';
-import getThemeFilterTermFromString from 'state/selectors/get-theme-filter-term-from-string';
+import { getThemeFilterTermFromString } from 'state/themes/selectors';
 
 import 'state/themes/init';
